requirements.txt
requests
pydantic
pandas
psycopg2-binary